requirements.txt
fastapi[standard]
websockets
python-dotenv